requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.6.4
openai==1.44.1
python-dotenv==1.0.0